final/requirements.txt
streamlit
flask
gunicorn
google-genai
feedparser
requests
python-dotenv
